move list struct (and CRUD functions) to separate file list.ex #68
