--- lib/app/ctx/list.ex
@@ -1,21 +1,0 @@
-defmodule App.Ctx.List do
-  use Ecto.Schema
-  import Ecto.Changeset
-
-  schema "lists" do
-    field :title, :string
-    # the owner of the list
-    field :person_id, :id
-    field :status, :id
-    field :tag, :id
-
-    timestamps()
-  end
-
-  @doc false
-  def changeset(list, attrs) do
-    list
-    |> cast(attrs, [:title])
-    |> validate_required([:title])
-  end
-end
