+++ lib/app/list.ex
@@ -1,0 +1,117 @@
+defmodule App.List do
+  use Ecto.Schema
+  import Ecto.Changeset
+  import Ecto.Query, warn: false
+  alias App.Repo
+  alias __MODULE__
+
+  schema "lists" do
+    field :title, :string
+    field :person_id, :id
+    field :status, :id
+    field :tag, :id
+
+    timestamps()
+  end
+
+  @doc false
+  def changeset(list, attrs) do
+    list
+    |> cast(attrs, [:title])
+    |> validate_required([:title])
+  end
+
+  @doc """
+  Returns the list of lists.
+
+  ## Examples
+
+      iex> list_lists()
+      [%List{}, ...]
+
+  """
+  def list_lists do
+    Repo.all(List)
+  end
+
+  @doc """
+  Gets a single list.
+
+  Raises `Ecto.NoResultsError` if the List does not exist.
+
+  ## Examples
+
+      iex> get_list!(123)
+      %List{}
+
+      iex> get_list!(456)
+      ** (Ecto.NoResultsError)
+
+  """
+  def get_list!(id), do: Repo.get!(List, id)
+
+  @doc """
+  Creates a list.
+
+  ## Examples
+
+      iex> create_list(%{field: value})
+      {:ok, %List{}}
+
+      iex> create_list(%{field: bad_value})
+      {:error, %Ecto.Changeset{}}
+
+  """
+  def create_list(attrs \\ %{}) do
+    %List{}
+    |> List.changeset(attrs)
+    |> Repo.insert()
+  end
+
+  @doc """
+  Updates a list.
+
+  ## Examples
+
+      iex> update_list(list, %{field: new_value})
+      {:ok, %List{}}
+
+      iex> update_list(list, %{field: bad_value})
+      {:error, %Ecto.Changeset{}}
+
+  """
+  def update_list(%List{} = list, attrs) do
+    list
+    |> List.changeset(attrs)
+    |> Repo.update()
+  end
+
+  @doc """
+  Deletes a List.
+
+  ## Examples
+
+      iex> delete_list(list)
+      {:ok, %List{}}
+
+      iex> delete_list(list)
+      {:error, %Ecto.Changeset{}}
+
+  """
+  def delete_list(%List{} = list) do
+    Repo.delete(list)
+  end
+
+  @doc """
+  Returns an `%Ecto.Changeset{}` for tracking list changes.
+
+  ## Examples
+
+      iex> change_list(list)
+      %Ecto.Changeset{source: %List{}}
+
+  """
+  def change_list(%List{} = list) do
+    List.changeset(list, %{})
+  end
+end
