+++ test/app/list_test.exs
@@ -1,0 +1,64 @@
+defmodule App.ListTest do
+  use App.DataCase
+  alias App.List
+
+  describe "lists" do
+    @valid_attrs %{title: "some title"}
+    @update_attrs %{title: "some updated title"}
+    @invalid_attrs %{title: nil}
+
+    def list_fixture(attrs \\ %{}) do
+      {:ok, list} =
+        attrs
+        |> Enum.into(@valid_attrs)
+        |> List.create_list()
+
+      list
+    end
+
+    test "list_lists/0 returns all lists" do
+      list = list_fixture()
+      assert List.list_lists() == [list]
+    end
+
+    test "get_list!/1 returns the list with given id" do
+      list = list_fixture()
+      assert List.get_list!(list.id) == list
+    end
+
+    test "create_list/1 with valid data creates a list" do
+      assert {:ok, %List{} = list} = List.create_list(@valid_attrs)
+      assert list.title == "some title"
+    end
+
+    test "create_list/1 with invalid data returns error changeset" do
+      assert {:error, %Ecto.Changeset{}} = List.create_list(@invalid_attrs)
+    end
+
+    test "update_list/2 with valid data updates the list" do
+      list = list_fixture()
+      assert {:ok, %List{} = list} = List.update_list(list, @update_attrs)
+      assert list.title == "some updated title"
+    end
+
+    test "update_list/2 with invalid data returns error changeset" do
+      list = list_fixture()
+
+      assert {:error, %Ecto.Changeset{}} =
+               List.update_list(list, @invalid_attrs)
+
+      assert list == List.get_list!(list.id)
+    end
+
+    test "delete_list/1 deletes the list" do
+      list = list_fixture()
+      assert {:ok, %List{}} = List.delete_list(list)
+      assert_raise Ecto.NoResultsError, fn -> List.get_list!(list.id) end
+    end
+
+    test "change_list/1 returns a list changeset" do
+      list = list_fixture()
+      assert %Ecto.Changeset{} = List.change_list(list)
+    end
+  end
+end
